requirements.txt
requests==2.31.0
pandas==2.1.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
apscheduler==3.10.4
python-dotenv==1.0.0
streamlit==1.31.0
plotly==5.18.0
pytest==7.4.4
pytest-mock==3.12.0
